Change include to src, refactor Makefile and apply format
