+++ src/dnaya/version.hpp.in
